Migrate environment-lifecycle
