+++ web/html/src/manager/content-management/shared/components/panels/environment-lifecycle/environment.utils.ts
@@ -3,21 +3,20 @@ export function mapAddEnvironmentRequest(environment, environments, projectId) {
     ...environment,
     projectLabel: projectId,
     label: environment.label,
-    name: environment.name
-  }
-
+    name: environment.name,
+  };
 
   if (!environmentRequest.predecessorLabel) {
-    const lastEnvironment = environments[environments.length - 1]
+    const lastEnvironment = environments[environments.length - 1];
     if (lastEnvironment) {
-      environmentRequest.predecessorLabel = environments[environments.length - 1].label
+      environmentRequest.predecessorLabel = environments[environments.length - 1].label;
     } else {
       delete environmentRequest.predecessorLabel;
     }
   } else {
-    const envIndex = environments.findIndex((env) => env.label === environmentRequest.predecessorLabel);
+    const envIndex = environments.findIndex(env => env.label === environmentRequest.predecessorLabel);
     if (envIndex > 0) {
-      environmentRequest.predecessorLabel = environments[envIndex - 1].label
+      environmentRequest.predecessorLabel = environments[envIndex - 1].label;
     } else {
       delete environmentRequest.predecessorLabel;
     }
@@ -30,12 +29,9 @@ export function mapUpdateEnvironmentRequest(environment, projectId) {
   let environmentRequest = {
     ...environment,
     projectLabel: projectId,
-  }
+  };
 
   return environmentRequest;
 }
 
-
-export function addEnvironmentAction() {
-
-}
+export function addEnvironmentAction() {}
